Rename docstringParser
